+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.109.0] - Not released
+### Fixed
+- Ignore minor exif-errors while sanitizing images.
+- Skip files which can not be sanitized because if errors after and report the errors.
+- In the case of registration with an already taken email, return 'This email
+  address is already in use' message instead of obscure 'Invalid email'.
+
 ## [1.108.1] - 2022-05-03
 ### Changed
 - Temporary turned off the batch attachments sanitize job handler